slight optimizations, improved bounce mechanic
before the ball bounced in such a way that its speed in the y direction would always be an integer mutilple of the player speed, meaning it was very simple for the ball to get back to 0 y speed which didnt feel right.

also replaced all of the x/2 to x*0.5 because it is a very simple optimization.